Begun implementation on brand new island generation system. This also introduces a new branch called "experimental". All content on here that is pushed will be _very_ unstable and might never be added to the final product. Feel free to test these commits, but please do not report bugs on them until they are pushed to master.
On the technical side of things, islands are now broken up into "IslandSectors", which will allow us to develop gameplay surrounding specific islands. This means in our code, we will be able to do lookups to nearby islands for things like Aerwhale rides to a chosen island, etc.

The actual island generation right now is VERY crude and unpolished, and has a lot of rough noise to it right now. Bare with me while I improve things.