Implemented new Baked Models for glowing or "fullbright" textures. Applied these models to the following blocks:
Ambrosium Ore
Labyrinth Lightstone
Labyrinth Base
Labyrinth Glowing Pillar

The Pillar model does not yet rotate visually, but will very soon. Haven't figured that one out yet.

Ambrosium Ore no longer emits light. This is a test to see if it improves performance on the server. If it does not, we will bring it back soon.