Snow mixing with grass.
My sugestion:
aether_grass_side_snowed.png + java.
Three images are worth more than 3000 words:
http://pt-br.tinypic.com/view.php?pic=359ahpl&s=9#.WBlxrIWcF9A
http://tinypic.com/view.php?pic=214v4n9&s=9
http://tinypic.com/view.php?pic=t9w5tc&s=9#.WBlyQYWcF9A

blighted and frostroot grass are different from normal minecraft and so
I did not do for them (I am not an artist).

There is snow that doesn't cover grass.
Please accept, Theoretically it is not a failure, but for me it is.