fix: Fixed interaction issues with Gravitite Tools, closes #372, #482, #493, #475
This makes Gravitite Tools more lenient towards which blocks can be picked up. If this proves to be problematic, we'll revert it back to the older more limited set of blocks.

Additionally, moving blocks will now try to pop out of the ground if they are impossible to place. This may lead to erratic looking behavior, but should be safer.