performance: Update OpenSimplexNoise with faster implementation
Noise sampling is around 50% faster than the previous implementation
and generates values in the range of [-1,1].

This updated generator does not produce identical results to the
previous implementation given the same inputs, but will produce the
same style of noise.

Our island generators need to be updated for this newer implementation.