change: New lighting system
This is a COMPLETE OVERHAUL of vanilla's lighting engine which brings
significant performance improvements in the Aether dimension and
various bug fixes for lighting bugs (i.e. dark spots in the world) for
the entire game.

These changes do not currently work outside of a de-obfuscated
(see: development) workspace. More work is needed to enable transformation
of obfuscated classes.

Special thanks to the work of PhiPro95 and Mathe172 towards identifying
and providing solutions to these issues. Their work has not made it
upstream into Minecraft Forge, which has led to us adopting their
changes directly as an embedded coremod in the Aether mod.

Some work is still needed to enable saving/reading of pending light
updates, which is leading to a few small bugs currently.