performance: Update OrbisAPI, significantly improve ore generation time
Ore generation in the Aether suffers from a problem that vanilla terrain
is mostly free from, being that there is a lot of air space beneath
and above islands where ores will try to generate. This changes the ore
generation system to take into account the min/max Y coordinates at the
position it is trying to generate at and adjust the ratios accordingly.

This provides an almost 50% reduction in ore generation time while
producing similar results. Theoretically, the same number of ores should
generate (and possibly more since fewer will fail in the air), but it
hasn't been thoroughly tested.