feature: Glactrix base implementation In current state, glactrix is only spawned from egg because it does not have textures or animations.
Glactrix will hide under shell and become invulnerable if it spots player.
If a player topples a Glactrix (by hurting it), it is shear-able.
If a Glactrix does not have a shell, it will attempt to run from the player,
and be vulnerable to attacks.
Glactrix will "regenerate" ice crystal's at night, after a set period of
time from when it was sheared.

Removed:
Icehog

Known issues:
Glactrix does not have a texture or animations.
Glactrix uses Taegore sound effects, and does not have a sound when player
attacks it while hiding.