performance: Implement new cave generation system
Islands now cache node data for caves in order to prevent excessive
overhead.

The logic to dig blocks and tunnel through caves has been optimized as
much as possible for the time being.

The cave generator performs a simple check to skip nodes outside the
filled portions of chunks, allowing us to skip a lot of nodes in air
sections of chunks.