refactor: Initial build script update
This is a **research project** piloted by Angeline.

In layman's terms, this branch is not a primary development focus. We're
making no ETAs as to when you can play the Aether II in Minecraft 1.13+
or even if we will pursue updating to the next version for the
foreseeable future.

Today, this will not compile. If you attempt to do so, you will waste
both your own and your computer's time senselessly. It will likely be
weeks before this will even manage to make it to the title screen, if we
get that far.

---

We're disabling Phosphor temporarily since the nature of it
significantly complicates the situation at hand. Research is needed as
to how ASM and ATs are handled in Forge for Minecraft 1.13+.