fix: Status Effects DISABLED     - disabled status effects for balance issues.
Note to Devs: Just commented out were they are registered to entities.
As far as I can tell there shouldn't be any issue since anytime an effect
is applied it should always check whether the capability is there. To
re-enable all you need to do is revert this commit/ uncomment the few lines
of code.