change: Significant changes to player module API
Player modules are now serialized in name-value pairs to disk rather
than as a sequential list of NBT data keyed by registration order. This
resolves a ticking time bomb that could cause complete corruption or
loss of all player module data when registration order changes.

Additionally, player modules are no longer defined as fields in the
PlayerAether class, and instead are accessed by a single method which
takes the class of the module to return a matching instance. Doing
so allows mods to add their own modules and reduces the amount code of
code to add new modules significantly.

This change breaks backwards comparability, and all old player data
for the Aether will be lost. Going forward, only modules which implement
the Serializable interface will be saved to disk.