refactor: PlayerAether#getPlayer can no longer return null or accept non-player entities
If the specified player entity does not have our capability, we will
instead throw an exception (which will likely crash the game.)

Player entities should ALWAYS have our capabilities, and null-checking
all over the place to silently return from methods causes many
issues when the capability is missing. The capability missing should
be treated as a critical error all the time.

The signature of the method has also changed to only accept player
entities. This ensures that programmers take the additional precaution
to ensure that the entity is a player.